feat: map to Edr chain type
