+++ v-next/hardhat/src/internal/edr/chain-type.ts
@@ -1,12 +1,18 @@
 import type { ChainType } from "../../types/network.js";
 
 import {
+  OP_CHAIN_TYPE as EDR_OP_CHAIN_TYPE,
+  L1_CHAIN_TYPE as EDR_L1_CHAIN_TYPE,
+  GENERIC_CHAIN_TYPE as EDR_GENERIC_CHAIN_TYPE,
+} from "@ignored/edr-optimism";
+
+import {
   GENERIC_CHAIN_TYPE,
   L1_CHAIN_TYPE,
   OPTIMISM_CHAIN_TYPE,
 } from "../constants.js";
 
-export function isEdrSupportedChainType(
+export function isSupportedChainType(
   chainType: unknown,
 ): chainType is ChainType {
   return (
@@ -14,4 +20,16 @@ export function isEdrSupportedChainType(
     chainType === L1_CHAIN_TYPE ||
     chainType === OPTIMISM_CHAIN_TYPE
   );
+}
+
+export function hardhatChainTypeToEdrChainType(chainType: ChainType): string {
+  if (chainType === OPTIMISM_CHAIN_TYPE) {
+    return EDR_OP_CHAIN_TYPE;
+  }
+
+  if (chainType === L1_CHAIN_TYPE) {
+    return EDR_L1_CHAIN_TYPE;
+  }
+
+  return EDR_GENERIC_CHAIN_TYPE;
 }
